刷新hotdomain

+++ domain-service/domain-service-client/src/main/java/com/gitgle/response/HotDomain.java
@@ -9,5 +9,5 @@ public class HotDomain implements Serializable {
 
     private String domain;
 
-    private Integer developerTotal;
+    private Long developerTotal;
 }
